move GeometryFilterProvider to sample
