+++ Samples/Mapsui.Samples.Common/Maps/WFS/WfsGeometryFilterSample.cs
@@ -1,4 +1,6 @@
-﻿using Mapsui.Extensions;
+﻿using System;
+using System.Collections.Generic;
+using Mapsui.Extensions;
 using Mapsui.Layers;
 using Mapsui.Logging;
 using Mapsui.Providers.Wfs;
@@ -7,7 +9,6 @@ using Mapsui.Tiling;
 using Mapsui.Widgets.InfoWidgets;
 using System.Net;
 using System.Threading.Tasks;
-using Mapsui.Nts.Providers;
 using Mapsui.Providers;
 
 #pragma warning disable IDISP001 // Dispose created
@@ -92,5 +93,42 @@ public class WfsGeometryFilterSample : ISample
                 LabelColumn = labelField
             }
         };
+    }
+    
+    private class GeometryFilterProvider(IProvider provider, Func<IFeature, bool> filter) : IProvider, IProviderExtended
+    {
+        private DateTime _lastUpdate;
+        private IEnumerable<IFeature> _current = [];
+        public int Id { get; } = BaseLayer.NextId();
+
+        public string? CRS
+        {
+            get => provider.CRS;
+            set => provider.CRS = value;
+        }
+
+        public async Task<IEnumerable<IFeature>> GetFeaturesAsync(FetchInfo fetchInfo)
+        {
+            if (DateTime.Now - _lastUpdate < TimeSpan.FromSeconds(20))
+            {
+                return _current;
+            }
+
+            _lastUpdate = DateTime.Now;
+            _current = IterateFeatures(await provider.GetFeaturesAsync(fetchInfo));
+            return _current;
+        }
+
+        private IEnumerable<IFeature> IterateFeatures(IEnumerable<IFeature> features)
+        {
+            foreach (var feature in features)
+                if (filter(feature))
+                    yield return feature;
+        }
+
+        public MRect? GetExtent()
+        {
+            return provider.GetExtent();
+        }
     }
 }
